requirements.txt
# Core dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.21
python-dotenv==1.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.1.0
tqdm==4.66.1

# Database and Vector Store
alembic==1.12.0
pymilvus==2.4.12

# Machine Learning
scikit-learn==1.6.0
joblib==1.4.2

# NLP and Chatbot
langchain==0.3.13
langchain-community==0.0.24
langchain-core==0.3.28
langchain-ollama==0.2.2

# Text Processing
nltk==3.8.1
tenacity==8.2.3

# Utilities
python-dateutil==2.8.2
gunicorn==21.2.0